MFH:
Normalize the list of packages to be added, and particularly removed,
occasionally pkg_delete <x> <x> will cause a false-positive failure
for a portbuild.  Further work can be done here to get a smart list of
packages so that we don't try to add packages that have already been
installed, important for ports with large numbers of dependencies.

Reported by:	pav