MFH:
Start a contrib directory to include user-contributed enhancements to
Tinderbox.  This will primarily be a Hooks repository.

Add a new Hook example for fixing the KSE problem in older Jails when
built on a -CURRENT build host.

Add a contributed Hook for setting ports as jobs safe based on the
maintainer.